backend/requirements.txt
fastapi
uvicorn
firebase-admin
google-generativeai
geopy
geolib
pydantic
